This should build OK on Linux, but there are some known problems on other OSs that I haven't fixed yet.
PS, I'd strongly encourage people to do a static server build as well
as a loader build before committing (#define DoLoadableServer NO).

3647. XFree86-Bigfont extension that provides efficient client access to
      the complete glyph metrics, using shared memory where possible/available
      or transmitting the data in a compressed format otherwise (#3626,
      Bruno Haible).
3646. Update locale support for Celtic languages (Welsh, Irish, Scots, Manx
      and Cornish), and include a compose file for ISO-8859-14 (#3612,
      Alastair McKinstry).
3644. The new extension init code didn't handle server resets (#3628,
      Ebgert Eich).
3643. xf86GetPciSizeFromOS() didn't close the file it opened (#3627,
      Loic Grenie).
3642. Export frexp() and ldexp() to modules (#3624, 3631, Juliusz Chroboczek).
3641. Fixes for CTEXT handling in Xlib:
       - rewrite CTEXT from/to converters for standard compliance
       - CTEXT parser now understands UTF-8 (and other DOC) sequences and
         performs only CTEXT sequence recognition
       - detailed charset information parsing moved to the procedure that
         adds 'charset object'
       - Some actions for preparing 'charset object' moved to lcCharSet.c
         (where it can be done easily)
       - XlcAddCT call added to XLC_LOCALE file parser to allow recognition
         of charsets that are described only in those files (ie, not
         in internal Xlib tables)
      (#3624, Ivan Pascal).
3640. Convert the nv driver to use MMIO to access the VGA registers
      instead of normal I/O ports.  Also riva_hw is updated to the
      same version as is in 3.3.6 (#3620, Jarno Paananen).
3639. DRI and Mesa update (#3619, Brian Paul, Precision Insight).
3638. Fix a problem with the tdfx driver that can result in a lockup
      (#3618, Nicholas Brealey).
3627. Fix some PPC inline assembler in compiler.h (#3617, Kostas Gewrgiou).
3626. Add support for the S3 Trio3D/2x and Trio3D chips to the s3virge
      driver (#3616, 3623, Harald Koenig).
3625. Fix some build problems on Solaris 8 with gcc (#3615, Richard Coley).
3624. Fix an Xserver core dump that can happen when xdmcp-related command
      line options have missing arguments (#3614, Harald Koenig).
3623. Xterm patch #125 (#3611, Thomas Dickey).
3622. Add a Type1 version of the cursor font (#3610, Juliusz Chroboczek).

+++ extras/Mesa/src/shade_tmp.h
@@ -1,4 +1,4 @@
-/* $Id: shade_tmp.h,v 1.1 1999/12/14 01:31:50 robin Exp $ */
+/* $Id: shade_tmp.h,v 1.2 2000/02/08 17:17:35 dawes Exp $ */
 
 /*
  * Mesa 3-D graphics library
@@ -135,7 +135,7 @@ static void TAG(shade_rgba_spec)( struct vertex_buffer *VB )
 	       SUB_3V(VP, light->Position, vertex);
 	    }
 
-	    d = LEN_3FV( VP );
+	    d = (GLfloat) LEN_3FV( VP );
 
 	    if (d > 1e-6) {
 	       GLfloat invd = 1.0F / d;
@@ -157,7 +157,7 @@ static void TAG(shade_rgba_spec)( struct vertex_buffer *VB )
 	       else {
 		  double x = PV_dot_dir * (EXP_TABLE_SIZE-1);
 		  int k = (int) x;
-		  GLfloat spot = (light->SpotExpTable[k][0]
+		  GLfloat spot = (GLfloat) (light->SpotExpTable[k][0]
 				  + (x-k)*light->SpotExpTable[k][1]);
 		  attenuation *= spot;
 	       }
@@ -234,7 +234,7 @@ static void TAG(shade_rgba_spec)( struct vertex_buffer *VB )
 	    }
 	    
 	    if (n_dot_h>1.0) {
-	       spec_coef = pow( n_dot_h, tab->shininess );
+	       spec_coef = (GLfloat) pow( n_dot_h, tab->shininess );
 	    } else 
 	       GET_SHINE_TAB_ENTRY( tab, n_dot_h, spec_coef );
 	    
@@ -733,7 +733,7 @@ static void TAG(shade_ci)( struct vertex_buffer *VB )
 	       SUB_3V(VP, light->Position, vertex);
 	    }
 
-	    d = (GLfloat) LEN_3FV( VP );
+	    d = LEN_3FV( VP );
 	    if ( d > 1e-6) {
 	       GLfloat invd = 1.0F / d;
 	       SELF_SCALE_SCALAR_3V(VP, invd);
@@ -753,7 +753,7 @@ static void TAG(shade_ci)( struct vertex_buffer *VB )
 	       else {
 		  double x = PV_dot_dir * (EXP_TABLE_SIZE-1);
 		  int k = (int) x;
-		  GLfloat spot = (GLfloat) (light->SpotExpTable[k][0]
+		  GLfloat spot = (light->SpotExpTable[k][0]
 				  + (x-k)*light->SpotExpTable[k][1]);
 		  attenuation *= spot;
 	       }
@@ -816,7 +816,7 @@ static void TAG(shade_ci)( struct vertex_buffer *VB )
 	    }
 	    
 	    if (n_dot_h>1.0) {
-	       spec_coef = (GLfloat) pow( n_dot_h, tab->shininess );
+	       spec_coef = pow( n_dot_h, tab->shininess );
 	    } else {
 	       GET_SHINE_TAB_ENTRY( tab, n_dot_h, spec_coef);
 	    }
